# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de gestion des catégories",
        "description": "Documentation pour l'API de gestion des catégories de votre application.",
        "contact": {
            "email": "votre_email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reservations": {
            "post": {
                "summary": "Crée une nouvelle réservation de ticket et gère le stock.",
                "description": "Valide les données, verrouille le stock pour éviter les problèmes de concurrence, vérifie la disponibilité, crée la réservation, met à jour le stock et génère les instances de tickets. Utilise une transaction de base de données.",
                "operationId": "01e3d21dc330a31de025203bc142ffa3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ticket_id",
                                    "quantite"
                                ],
                                "properties": {
                                    "ticket_id": {
                                        "description": "ID du ticket à réserver (doit exister dans la table 'tickets').",
                                        "type": "integer"
                                    },
                                    "quantite": {
                                        "description": "Quantité de tickets à réserver.",
                                        "type": "integer",
                                        "format": "int32",
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Réservation créée avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Réservation créée avec succès. 5 tickets générés."
                                        },
                                        "reservation_id": {
                                            "description": "ID de la nouvelle réservation.",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Quantité de tickets non disponible (stock insuffisant).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Quantité de tickets non disponible (stock restant: 2)."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation des données fournies.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur lors du traitement de la transaction.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors du traitement de la réservation."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/ticket-by-event/{evenementId}": {
            "get": {
                "summary": "Récupérer toutes les instances de tickets pour un événement donné",
                "operationId": "272235fd8b00e07ea9a60dea253b4e69",
                "parameters": [
                    {
                        "name": "evenementId",
                        "in": "path",
                        "description": "ID de l'événement",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des instances de tickets pour l'événement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucune instance de ticket trouvée pour cet événement"
                    }
                }
            }
        },
        "/api/ticket-by-event-and-payment-status/{evenementId}/{isPaid}": {
            "get": {
                "summary": "Récupérer les instances de tickets par événement et statut de paiement",
                "description": "Récupère les instances de tickets (Ticketinstance) pour un événement donné, filtrées par leur statut de paiement (1 pour Payé, 0 pour Non Payé).",
                "operationId": "3374007582de6df418f3eb73ee56d1eb",
                "parameters": [
                    {
                        "name": "evenementId",
                        "in": "path",
                        "description": "ID de l'événement",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "isPaid",
                        "in": "path",
                        "description": "Statut de paiement : 1 pour Payé, 0 pour Non Payé",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des instances de tickets correspondant aux critères",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Statut de paiement invalide (doit être 0 ou 1)"
                    },
                    "404": {
                        "description": "Aucune instance de ticket trouvée pour cet événement et ce statut"
                    }
                }
            }
        },
...
    },
...
}